+++ src/app/services/spreadsheet.service.ts
@@ -27,6 +27,7 @@ import {
   createSlug,
   getDateFromString,
   getFullUrl,
+  slugify,
 } from 'src/app/shared/utils';
 import { environment } from 'src/environments/environment';
 
@@ -375,6 +376,10 @@ export class SpreadsheetService {
         row,
         colMap.get(OfferCol.moreInfo),
       ),
+      chapterName: SpreadsheetService.readCellValue(
+        row,
+        colMap.get(OfferCol.chapter),
+      ),
     };
   }
 
@@ -398,7 +403,15 @@ export class SpreadsheetService {
           .map((row: string[]) =>
             this.convertOfferRowToOfferObject(row, offerColumnMap),
           )
-          .filter((offer: Offer): boolean => offer.offerVisible);
+          .filter((offer: Offer): boolean => offer.offerVisible)
+          .map((offer: Offer) => {
+            offer.chapterSlug = createSlug(
+              null,
+              slugify(offer.chapterName),
+              SlugPrefix.chapter,
+            );
+            return offer;
+          });
       })
       .catch((error) => {
         this.loggingService.logException(error);